refactor: Close memory architecture boundaries
This patch establishes strict layer boundaries for the memory subsystem as
dictated by the memory gap closure plan (Phase 0).

Changes:
- Established target directories: `pmm/`, `vm/objects/`, `vm/aspace/`, `vm/fault/`, `dma/`, `iommu/`, `hal/mmu/`.
- Moved existing `.c` files to respect the new structure, updating CMake build definitions.
- Normalized public headers (`pmm.h`, `vm_object.h`, `aspace.h`, `fault.h`, `dma.h`, `hal_pt.h`, `hal_tlb.h`, `hal_iommu.h`), injecting strict ownership contract rules, locking expectations, and profile applicability (`PROFILE_MMU_FULL`, `PROFILE_MMU_LITE`, `PROFILE_MPU_ONLY`).
- Created isolated internal headers per domain (`pmm_internal.h`, `aspace_internal.h`, `vm_object_internal.h`, `mmu_internal.h`) to hide specific logic from public contracts.
- Removed legacy cross-layer shortcuts, replacing instances of `hal_vmm_get_mapping`, `hal_vmm_update_mapping`, and `active_mmu` direct struct access with the robust architecture-neutral contracts `active_hal_pt` and `active_hal_tlb`.
- Produced a summary audit document detailing the newly enforced boundaries.
