+++ kernel/include/mm.h
@@ -119,3 +119,4 @@ void vmm_process_urpc_messages(void);
 #endif // BHARAT_MM_H
 
 void tlb_shootdown(address_space_t *as, virt_addr_t vaddr);
+#define PAGE_EXEC 0x400
