+++ kernel/src/hal/mmu/README.md
@@ -1,0 +1,4 @@
+# HAL MMU Layer
+- Arch implementation only, no upward policy logic.
+- May call: None.
+- Forbidden: VM policy/fault logic.
