+++ kernel/src/hal/mmu/mmu_internal.h
@@ -1,0 +1,11 @@
+/*
+ * MMU Internal Logic
+ * Not to be exposed to generic VM code
+ */
+#ifndef MMU_INTERNAL_H
+#define MMU_INTERNAL_H
+
+#include "../../../include/hal/hal_pt.h"
+#include "../../../include/hal/hal_tlb.h"
+
+#endif
