+++ kernel/src/mm/iommu/README.md
@@ -1,0 +1,4 @@
+# IOMMU Layer
+- Owns HAL-facing / domain-facing, isolated from generic VM policy.
+- May call: HAL IOMMU.
+- Forbidden: Generic VM policy.
