+++ kernel/src/mm/pmm/README.md
@@ -1,0 +1,4 @@
+# PMM Layer
+- Owns physical frame discovery, zones, buddy/contiguous alloc, refcounts, pinned/reserved classes.
+- May call: None
+- Forbidden: VM policy, fault layers, arch internals.
