+++ kernel/src/mm/pmm/pmm_internal.h
@@ -1,0 +1,11 @@
+/*
+ * PMM Internal Logic
+ * Not to be exposed to vm/aspace or hal/
+ */
+#ifndef PMM_INTERNAL_H
+#define PMM_INTERNAL_H
+
+#include "../../include/mm/pmm.h"
+#include "../../include/spinlock.h"
+
+#endif
