+++ kernel/src/mm/vm/aspace/aspace_internal.h
@@ -1,0 +1,10 @@
+/*
+ * ASpace Internal Logic
+ * Not to be exposed to pmm/ or hal/
+ */
+#ifndef ASPACE_INTERNAL_H
+#define ASPACE_INTERNAL_H
+
+#include "../../../include/mm/aspace.h"
+
+#endif
