+++ kernel/src/mm/vm/objects/vm_object_internal.h
@@ -1,0 +1,10 @@
+/*
+ * VM Object Internal Logic
+ * Not to be exposed to aspace/ or pmm/
+ */
+#ifndef VM_OBJECT_INTERNAL_H
+#define VM_OBJECT_INTERNAL_H
+
+#include "../../../include/mm/vm_object.h"
+
+#endif
